AI-RCM/src/payment_posting.py
```python
import pandas as pd
from models.payment_posting_model import PaymentPostingModel
from utils import load_data, save_data

class PaymentPosting:
    def __init__(self):
        self.model = PaymentPostingModel()
        self.payment_data = load_data('data/payment_data.py')

    def match_payments_to_claims(self):
        matched_payments = self.model.match_payments_to_claims(self.payment_data)
        return matched_payments

    def automate_payment_posting(self):
        automated_payments = self.model.automate_payment_posting(self.payment_data)
        return automated_payments

    def post_payments(self):
        matched_payments = self.match_payments_to_claims()
        automated_payments = self.automate_payment_posting()

        posted_payments = pd.concat([matched_payments, automated_payments], axis=1)
        save_data(posted_payments, 'data/posted_payments.py')

if __name__ == "__main__":
    payment_posting = PaymentPosting()
    payment_posting.post_payments()
```